+++ tier1/lzss.cpp
@@ -55,7 +55,7 @@ void CLZSS::BuildHash( const unsigned char *pData )
 	lzss_list_t *pList;
 	lzss_node_t *pTarget;
 
-	int targetindex = (unsigned int)pData & ( m_nWindowSize - 1 );
+	int targetindex = (size_t)pData & ( m_nWindowSize - 1 );
 	pTarget = &m_pHashTarget[targetindex];
 	if ( pTarget->pData )
 	{
